:card_file_box: Add room titles and descriptions
Adding titles and descriptions for each square(room) of the maze, in addition to the already existing deployed set of coordinate data.  Maze data will produce the expected 4X4 maze that was requested and will remain unchanged to align with the user stories related to multiple players and their room positions.