:construction: Room model and implementation started
Created Room model and manager: end goal is for each Room instance to hold an array containing the usernames of all the players currently in it. On a player move(), API would send back the room which contained the updated players array to be displayed in FE display. This would also require extra function for the Users PUT request to remove their username when leaving a room, and adding it to the next room's player array on every move